:recycle: Refactoring changes thanks to PGA
